+++ hata/discord/guild/audit_logs/target_converters.py
@@ -129,5 +129,9 @@ def target_converter_application_command(entry):
             return parent.application_commands(target_id, None)
 
 
-def target_converter_auto_moderation(entry):
-    return None
+def target_converter_auto_moderation_rule(entry):
+    target_id = entry.target_id
+    if target_id:
+        parent = entry.parent
+        if (parent is not None):
+            return parent.auto_moderation_rules(target_id, None)