+++ hata/discord/scheduled_event/metadata.py
@@ -12,7 +12,7 @@ def try_get_scheduled_event_metadata_type_from_data(data):
     Parameters
     ----------
     data : `dict` of (`str`, `str`) items
-        Scheduled event metadata type.
+        Scheduled event metadata data.
     
     Returns
     -------